template compilator is done. run npm build
