+++ dev/compile_templates.js
@@ -1,29 +1,13 @@
 const fs = require('fs');
 const path = require('path');
+const pug = require('pug');
 
 const correctLaunchDir = '2021_2_SaberDevs';
-const devRoot = "dev/"
-const outRoot = "public/components"
-const templatesDir = "components/"
+const devRoot = "dev/";
+const templatesDir = "components/";
+const outDir = "public/components/";
+const componentSuffix = "Component";
 
-function transformTemplate(filename) {
-    // уберешь utf8 - будешь читать побайтово. data станет переменной типа fs.Buffer
-    fs.readFile(filename, 'utf8', (err, data) => {
-        if (err) return console.log(err);
-
-        const ext = filename.slice(filename.lastIndexOf('.') + 1);
-        // трогаем только js
-        if (ext != 'js') return;
-
-        const functionName = filename.slice(filename.lastIndexOf('/') + 1, filename.lastIndexOf('.'));
-        const result = data.replace(/function template/g, '\n\nexport default function ' + functionName);
-      
-        fs.writeFile(filename, result, 'utf8', (err) => {
-           if (err) return console.log(err);
-           console.log("\tФайл", functionName + '.' + ext, "преобразован");
-        });
-    })
-};
 
 
 console.log("\t===============================");
@@ -41,7 +25,45 @@ if (currentDir !== correctLaunchDir) {
 
 fs.readdir(devRoot + templatesDir, (err, files) => {
     if (err) return console.log(err);
-    files.forEach(file => {
-        transformTemplate(devRoot + templatesDir + file);
+    files.forEach(filename => {
+        // проверяем, что не директория, и что разрешение есть
+        const dotPos = filename.lastIndexOf('.');
+        if (dotPos == -1) return
+        // трогаем только pug
+        const ext = filename.slice(dotPos + 1);
+        if (ext != 'pug') return;
+
+        const functionName = filename.slice(filename.lastIndexOf('/') + 1, filename.lastIndexOf('.'));
+        // Compile a function
+        // Дебил, смотри доку внимательнее
+        // https://pugjs.org/api/reference.html
+        const compiledTemplate = pug.compileFileClient(devRoot + templatesDir + filename, {name: 'x(){};\n\nexport default function ' + functionName + componentSuffix});
+
+        fs.writeFile(outDir + filename.replace(/pug/, "js"), compiledTemplate, 'utf8', (err) => {
+            if (err) return console.log(err);
+            console.log("\Шаблон", filename, "преобразован в " + outDir + functionName + ".js");
+        });
     });
 });
+
+
+
+// Не нужна. Пускай лежит. Позволяет исправлять уже сгенерированные pug'ом js-файлы
+function transformTemplate(filename) {
+    // уберешь utf8 - будешь читать побайтово. data станет переменной типа fs.Buffer
+    fs.readFile(filename, 'utf8', (err, data) => {
+        if (err) return console.log(err);
+
+        const ext = filename.slice(filename.lastIndexOf('.') + 1);
+        // трогаем только js
+        if (ext != 'js') return;
+
+        const functionName = filename.slice(filename.lastIndexOf('/') + 1, filename.lastIndexOf('.'));
+        const result = data.replace(/function template/g, '\n\nexport default function ' + functionName);
+      
+        fs.writeFile(outDir + filename, result, 'utf8', (err) => {
+           if (err) return console.log(err);
+           console.log("\tФайл", functionName + '.' + ext, "преобразован");
+        });
+    })
+};
