+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "node ./dev/compile_templates.js"
+    "build": "pug node ./dev/compile_templates.js",
+    "start": "node server/server.js"
   },
   "keywords": [],
   "author": "Denis, Svetlana, Yuriy, Alexey",
