+++ public/components/profile npm pug.js
@@ -1,0 +1,25 @@
+function pug_rethrow(n,e,r,t){
+    if(!(n instanceof Error))
+        throw n;
+    if(!("undefined"==typeof window&&e||t))
+        throw n.message+=" on line "+r,n;
+    try{
+        t=t||require("fs").readFileSync(e,"utf8")
+    }
+    catch(e){
+        pug_rethrow(n,null,r)
+    }
+    var i=3,a=t.split("\n"),o=Math.max(r-i,0),h=Math.min(a.length,r+i),i=a.slice(o,h).map(function(n,e){var t=e+o+1;
+    return(t==r?"  > ":"    ")+t+"| "+n}).join("\n");throw n.path=e,n.message=(e||"Pug")+":"+r+"\n"+i+"\n\n"+n.message,n
+}
+function x(){};
+
+export default function profileComponent(locals) {
+    console.log("shit happens. locals = ", locals)
+    var pug_html = "", pug_mixins = {}, pug_interp;var pug_debug_filename, pug_debug_line;
+    try {
+        ;pug_debug_line = 1;
+        pug_html = pug_html + "\u003Cprofile class=\"pug\"\u003E\u003C\u002Fprofile\u003E";
+    } catch (err) {
+        pug_rethrow(err, pug_debug_filename, pug_debug_line);
+    };return pug_html;}
